+++ erpnext/accounts/doctype/gl_entry/gl_entry.py
@@ -32,7 +32,7 @@ class GLEntry(Document):
 		self.validate_dimensions_for_pl_and_bs()
 
 		if not self.flags.from_repost:
-			self.check_pl_account()
+			# self.check_pl_account()
 			self.validate_party()
 			self.validate_currency()
 
@@ -246,9 +246,18 @@ def update_outstanding_amt(account, party_type, party, against_voucher_type, aga
 			
 			for advance in ref_doc.get("advances"):
 				bal -= advance.allocated_amount
+			
+			documents = frappe.get_all("Payment Entry Reference", ["*"], filters = {"reference_name": against_voucher})
+
+			if len(documents) > 0:
+				for document in documents:
+					entry = frappe.get_doc("Payment Entry", document.parent)
+
+					if entry.docstatus == 1:
+						bal -= document.allocated_amount
 		
 			if bal < 0:
-				bal = 0
+				bal = 0			
 
 		ref_doc.db_set('outstanding_amount', bal)
 		ref_doc.set_status(update=True)
