+++ erpnext/stock/doctype/inventory_adjustment/inventory_adjustment.js
@@ -1,0 +1,8 @@
+// Copyright (c) 2022, Frappe Technologies Pvt. Ltd. and contributors
+// For license information, please see license.txt
+
+frappe.ui.form.on('Inventory Adjustment', {
+	// refresh: function(frm) {
+
+	// }
+});
