+++ erpnext/stock/doctype/inventory_adjustment/test_inventory_adjustment.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2022, Frappe Technologies Pvt. Ltd. and Contributors
+# See license.txt
+from __future__ import unicode_literals
+
+# import frappe
+import unittest
+
+class TestInventoryAdjustment(unittest.TestCase):
+	pass
