+++ erpnext/stock/doctype/inventory_adjustment_items/inventory_adjustment_items.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2022, Frappe Technologies Pvt. Ltd. and contributors
+# For license information, please see license.txt
+
+from __future__ import unicode_literals
+# import frappe
+from frappe.model.document import Document
+
+class InventoryAdjustmentItems(Document):
+	pass
